phase2_weeklyPulse/requirements.txt
fastapi
uvicorn
python-dotenv
google-play-scraper
